Replacing INT_MAX by DBL_MAX

+++ include/Tracker.h
@@ -86,7 +86,7 @@ public:
         return max;
       });
       minR.setup([this]() {
-        double min = INT_MAX; 
+        double min = DBL_MAX; 
         for (const auto& b : barrels_) min = MIN(min, b.minR());
         for (const auto& e : endcaps_) min = MIN(min, e.minR());
         return min;
